requirements.txt
beautifulsoup4==4.14.3
langchain_anthropic==1.4.0
langchain_community==0.4.1
langchain_core==1.2.28
langchain_mcp_adapters==0.2.2
langchain_openai==1.1.12
langchain_upstage==0.7.7
langgraph==1.1.6
mcp==1.27.0
openai==2.31.0
pydantic==2.12.5
python-dotenv==1.2.2
pytz==2025.2
Requests==2.33.1
slack_sdk==3.41.0